Add solution for challenge 2 and challenge 3

+++ src/estimator.py
@@ -19,8 +19,8 @@ def estimator(data):
       'infectionsByRequestedTime': (data['reportedCases'] * 10) * (2 ** factor),
       'severeCasesByRequestedTime': int(.15 * (data['reportedCases'] * 10) * (2 ** factor)),
       'hospitalBedsByRequestedTime': int((data['totalHospitalBeds'] * .35) - (.15 * (data['reportedCases'] * 10) * (2 ** factor))),
-      'casesForICUByRequestedTime': .05 * ((data['reportedCases'] * 10) * (2 ** factor)),
-      'casesForVentilatorsByRequestedTime': 0.02 * ((data['reportedCases'] * 10) * (2 ** factor)),
+      'casesForICUByRequestedTime': int(.05 * ((data['reportedCases'] * 10) * (2 ** factor))),
+      'casesForVentilatorsByRequestedTime': int(0.02 * ((data['reportedCases'] * 10) * (2 ** factor))),
       'dollarsInFlight': int(((data['reportedCases'] * 10) * (2 ** factor) * .65 * 1.5) / 30)
     },
     'severeImpact': {
@@ -28,8 +28,8 @@ def estimator(data):
       'infectionsByRequestedTime': (data['reportedCases'] * 50) * (2 ** factor),
       'severeCasesByRequestedTime': int(.15 * (data['reportedCases'] * 50) * (2 ** factor)),
       'hospitalBedsByRequestedTime': int((data['totalHospitalBeds'] * .35) - (.15 * (data['reportedCases'] * 50) * (2 ** factor))),
-      'casesForICUByRequestedTime': .05 * ((data['reportedCases'] * 50) * (2 ** factor)),
-      'casesForVentilatorsByRequestedTime': 0.02 * ((data['reportedCases'] * 50) * (2 ** factor)),
+      'casesForICUByRequestedTime': int(.05 * ((data['reportedCases'] * 50) * (2 ** factor))),
+      'casesForVentilatorsByRequestedTime': int(0.02 * ((data['reportedCases'] * 50) * (2 ** factor))),
       'dollarsInFlight': int(((data['reportedCases'] * 50) * (2 ** factor) * .65 * 1.5) / 30)
     }
   }
